Announce calendar events
A known issue is that we're getting false alarms when an event is edited on the
device.

They are however timestamped with the time of the change, not the time of any
alarms, so what happens is that we don't find any matching events when looking
for them, and we don't say anything. We should treat this in a good way; logging
exceptions every time it happens saying that no events were found might not be
it.

Another known issue is that we announce declined events. We shouldn't do that.