Add a Privacy Policy as required by Google
As a side effect of adding this policy, stop sending app logs to
Crashlytics. Those logs are potentially sensitive, and retaining them
would make the privacy policy much more complex.

If people want to report when the app "feels" bad, they can do so using
the contact-the-developer functionality built into the app.